Split unneeded Core Settlement component
