+++ config/packages/_sylius.yaml
@@ -13,7 +13,6 @@ imports:
     - { resource: "./../../src/Component/Core/Common/Resources/config.yaml" }
     - { resource: "./../../src/Component/Core/Shop/Resources/config.yaml" }
     - { resource: "./../../src/Component/Core/Vendor/Resources/config.yaml" }
-    - { resource: "./../../src/Component/Core/Settlement/Resources/config.yaml" }
     - { resource: "./../../src/Component/Messaging/Resources/config.yaml" }
     - { resource: "./../../src/Component/Override/Resources/config.yaml" }
