+++ src/Component/Core/Admin/Resources/config.yaml
@@ -1,2 +1,3 @@
 imports:
     - { resource: 'grids/*.yaml' }
+    - { resource: 'ui/*.yaml' }
