+++ src/Component/Core/Admin/Resources/services/twig_extensions.xml
@@ -11,8 +11,8 @@ We are hiring developers from all over the world. Join us and start your new, ex
     <services>
         <defaults autowire="false" autoconfigure="false" />
 
-        <service class="BitBag\OpenMarketplace\Component\Core\Settlement\Twig\Extension\SettlementOrderCountExtension"
-                 id="bitbag.open_marketplace.component.core.settlement.twig.extension.settlement_order_count">
+        <service class="BitBag\OpenMarketplace\Component\Core\Admin\Twig\Extension\SettlementOrderCountExtension"
+                 id="bitbag.open_marketplace.component.core.admin.twig.extension.settlement_order_count">
             <argument type="service" id="sylius.repository.order" />
             <tag name="twig.extension" />
         </service>
