+++ src/Component/Core/Admin/Twig/Extension/SettlementOrderCountExtension.php
@@ -9,7 +9,7 @@
 
 declare(strict_types=1);
 
-namespace BitBag\OpenMarketplace\Component\Core\Settlement\Twig\Extension;
+namespace BitBag\OpenMarketplace\Component\Core\Admin\Twig\Extension;
 
 use BitBag\OpenMarketplace\Component\Order\Repository\OrderRepositoryInterface;
 use BitBag\OpenMarketplace\Component\Settlement\Entity\SettlementInterface;
