+++ src/Component/Core/Common/Controller/Resource/SettlementController.php
@@ -9,7 +9,7 @@
 
 declare(strict_types=1);
 
-namespace BitBag\OpenMarketplace\Component\Core\Settlement\Controller;
+namespace BitBag\OpenMarketplace\Component\Core\Common\Controller\Resource;
 
 use Sylius\Bundle\ResourceBundle\Controller\ResourceController;
 use Sylius\Component\Resource\ResourceActions;
