+++ src/Component/Core/Common/Resources/resources/settlement.yaml
@@ -7,4 +7,4 @@ sylius_resource:
         interface: BitBag\OpenMarketplace\Component\Settlement\Entity\SettlementInterface
         repository: BitBag\OpenMarketplace\Component\Settlement\Repository\SettlementRepository
         factory: BitBag\OpenMarketplace\Component\Settlement\Factory\SettlementFactory
-        controller: BitBag\OpenMarketplace\Component\Core\Settlement\Controller\SettlementController
+        controller: BitBag\OpenMarketplace\Component\Core\Common\Controller\Resource\SettlementController
