+++ src/Component/Core/Common/Resources/services/controllers.xml
@@ -63,8 +63,8 @@ We are hiring developers from all over the world. Join us and start your new, ex
             <argument type="service" id="bitbag.open_marketplace.component.vendor.context.vendor"/>
         </service>
 
-        <service class="BitBag\OpenMarketplace\Component\Core\Settlement\Controller\Action\ProfitWithdrawalAction"
-                 id="bitbag.open_marketplace.component.core.settlement.controller.action.profit_withdrawal"
+        <service class="BitBag\OpenMarketplace\Component\Core\Vendor\Controller\Settlement\ProfitWithdrawalAction"
+                 id="bitbag.open_marketplace.component.core.vendor.controller.settlement.profit_withdrawal"
         >
             <argument type="service">
                 <service class="Sylius\Component\Resource\Metadata\MetadataInterface">
