--- src/Component/Core/Settlement/Resources/config.yaml
@@ -1,2 +1,0 @@
-imports:
-    - { resource: 'ui/ui.yaml' }
