--- src/Component/Core/Settlement/Resources/services.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-
-<!--
-This file was created by developers working at BitBag
-Do you need more information about us and what we do? Visit our https://bitbag.io website!
-We are hiring developers from all over the world. Join us and start your new, exciting adventure and become part of us: https://bitbag.io/career
--->
-
-<container xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://symfony.com/schema/dic/services"
-           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-    <imports>
-        <import resource="services/*.xml"/>
-    </imports>
-    <services>
-        <defaults autowire="false" autoconfigure="false" />
-
-    </services>
-</container>
