--- src/Component/Core/Settlement/Resources/services/form.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-
-<!--
-This file was created by developers working at BitBag
-Do you need more information about us and what we do? Visit our https://bitbag.io website!
-We are hiring developers from all over the world. Join us and start your new, exciting adventure and become part of us: https://bitbag.io/career
--->
-
-<container xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://symfony.com/schema/dic/services"
-           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-    <parameters>
-        <parameter key="open_marketplace.profit_withdrawal.validation_groups" type="collection">
-            <parameter>sylius</parameter>
-        </parameter>
-    </parameters>
-    <services>
-        <defaults autowire="false" autoconfigure="false" />
-        <service class="BitBag\OpenMarketplace\Component\Core\Settlement\Form\ProfitWithdrawalType"
-                id="open_marketplace.component.core.settlement.form.profit_withdrawal_type"
-        >
-            <argument type="service" id="sylius.repository.channel" />
-            <argument type="service" id="request_stack" />
-            <tag name="form.type"/>
-        </service>
-    </services>
-</container>
